server/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
django-jsonview==2.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13